+++ lib/assets/javascripts/new-dashboard/store/modules/datasets/index.js
@@ -1,5 +1,5 @@
 import toObject from 'new-dashboard/utils/to-object';
-import Filters, { defaultParams as filtersDefaultParams } from 'new-dashboard/core/configuration/filters';
+import Filters from 'new-dashboard/core/configuration/filters';
 
 import * as VisualizationActions from '../../actions/visualizations';
 import * as VisualizationMutations from '../../mutations/visualizations';
@@ -54,7 +54,7 @@ const datasets = {
 
     setPagination (state, page = 1) {
       state.page = page;
-      state.numPages = Math.ceil(state.metadata.total_entries / filtersDefaultParams.per_page) || 1;
+      state.numPages = Math.ceil(state.metadata.total_entries / state.resultsPerPage) || 1;
     },
 
     setFilterType (state, filterType = 'mine') {
